feat: export function makeHierarchicNodes about sankey
